requirements.txt
azure-common==1.1.27
azure-core==1.15.0
azure-nspkg==3.0.2
azure-storage-blob==12.8.1
azure-storage-common==0.37.1
azure-storage-nspkg==3.1.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
Flask==1.1.2
futures==3.1.1
idna==2.10
install==1.3.4
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
msrest==0.6.21
numpy==1.20.3
oauthlib==3.1.1
Pillow==8.2.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1